[feat] 스코어 획득, 로드/세이브 기능 구현
- LobbyScene
-- 최대 스코어를 알려주는 UI 추가
- GameScene
-- 몬스터마다 자신의 점수를 가짐
-- 몬스터가 파괴되면 플레이어의 점수가 늘어남

+++ Assets/Scripts/Data/DataStructures/UserDataStructure.cs
@@ -3,6 +3,6 @@ namespace PFS.Data.DataStructures.userDataStructure
     struct UserDataStructure
     {
         public bool is_newbie;
-        public int max_score;
+        public ulong max_score;
     }
 }
