+++ Assets/Scripts/Data/StaticData/StaticUserData.cs
@@ -3,6 +3,6 @@ namespace PFS.Data.StaticData.staticUserData
     public static class StaticUserData
     {
         public static bool isNewbie = true;
-        public static int maxScore = 0;
+        public static ulong maxScore = 0;
     }
 }
